Clone dialog redesign, custom dropdowns, template slim-down
Frontend: two-panel New-clone dialog with rail tabs and left-label fields;
DropdownSelect replaces every native select; modal exit animation with
end-state hold; pinned dialog heights.
Template: drop uv/rustup/nvm, single-1080p (no monitor args), rsync+rclone,
comment cuts, gnome-patch moved under template/ with shell-03 renumbered
to shell-02.

+++ crates/clone-daemon/src/capture_pw.rs
@@ -376,7 +376,7 @@ fn build_buffers_pod() -> Result<Vec<u8>> {
                 flags: PropertyFlags::MANDATORY,
                 value: Value::Choice(spa::pod::ChoiceValue::Int(Choice(
                     ChoiceFlags::empty(),
-                    ChoiceEnum::Flags { default: data_types as i32, flags: vec![] },
+                    ChoiceEnum::Flags { default: data_types, flags: vec![] },
                 ))),
             },
         ],
